Change entities package
