+++ src/java/entities/Enrolled.java
@@ -1,4 +1,4 @@
-package entitis;
+package entities;
 
 import java.io.Serializable;
 import javax.persistence.EmbeddedId;
