+++ src/java/entities/EnrolledId.java
@@ -1,4 +1,4 @@
-package entitis;
+package entities;
 
 import java.io.Serializable;
 import java.util.Objects;
