+++ src/java/entities/Exam.java
@@ -3,16 +3,20 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
+import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.io.File;
 import java.io.Serializable;
 import java.util.Date;
+import java.util.Set;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import javax.persistence.Temporal;
 import javax.xml.bind.annotation.XmlRootElement;
 
 /**
@@ -20,95 +24,95 @@ import javax.xml.bind.annotation.XmlRootElement;
  * @author 2dam
  */
 @Entity
-@Table(name = "exercise", schema = "bytebuddiesbd")
+@Table(name = "exam", schema = "bytebuddiesbd")
 @XmlRootElement
-public class Exercise implements Serializable{
+public class Exam implements Serializable {
+
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Integer id;
-    private String name;
     private String description;
-    private LevelType levelType;
+    @Temporal(javax.persistence.TemporalType.DATE)
+    @JsonSerialize(as = Date.class)
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ssXXX")
+    private Date dateInit;
+    private Integer duration;
     private File file;
-    private File fileSolution;
-    private Date deadline;
-    private Integer hours;
-    private Unit unit;
-    
-    //Setters and Getters
+    private Subject subject;
+    private Set<Mark> marks;
 
+    //Setters and Getters
     public Integer getId() {
         return id;
     }
+
     public void setId(Integer id) {
         this.id = id;
     }
-    public String getName() {
-        return name;
-    }
-    public void setName(String name) {
-        this.name = name;
-    }
+
     public String getDescription() {
         return description;
     }
+
     public void setDescription(String description) {
         this.description = description;
     }
-    public LevelType getLevelType() {
-        return levelType;
+
+    public Date getDateInit() {
+        return dateInit;
     }
-    public void setLevelType(LevelType levelType) {
-        this.levelType = levelType;
+
+    public void setDateInit(Date dateInit) {
+        this.dateInit = dateInit;
     }
+
+    public Integer getDuration() {
+        return duration;
+    }
+
+    public void setDuration(Integer duration) {
+        this.duration = duration;
+    }
+
     public File getFile() {
         return file;
     }
+
     public void setFile(File file) {
         this.file = file;
     }
-    public File getFileSolution() {
-        return fileSolution;
+
+    public Subject getSubject() {
+        return subject;
     }
-    public void setFileSolution(File fileSolution) {
-        this.fileSolution = fileSolution;
+
+    public void setSubject(Subject subject) {
+        this.subject = subject;
     }
-    public Date getDeadline() {
-        return deadline;
+
+    public Set<Mark> getMarks() {
+        return marks;
     }
-    public void setDeadline(Date deadline) {
-        this.deadline = deadline;
+
+    public void setMarks(Set<Mark> marks) {
+        this.marks = marks;
     }
-    public Integer getHours() {
-        return hours;
-    }
-    public void setHours(Integer hours) {
-        this.hours = hours;
-    }
-    public Unit getUnit() {
-        return unit;
-    }
-    public void setUnit(Unit unit) {
-        this.unit = unit;
-    }
-    
-    //Constructors
-    public Exercise(Integer id, String name, String description, LevelType levelType, File file, File fileSolution, Date deadline, Integer hours, Unit unit) {
+
+    //Constructor
+    public Exam(Integer id, String description, Date dateInit, Integer duration, File file, Subject subject, Set<Mark> marks) {
         this.id = id;
-        this.name = name;
         this.description = description;
-        this.levelType = levelType;
+        this.dateInit = dateInit;
+        this.duration = duration;
         this.file = file;
-        this.fileSolution = fileSolution;
-        this.deadline = deadline;
-        this.hours = hours;
-        this.unit = unit;
+        this.subject = subject;
+        this.marks = marks;
     }
-    public Exercise() {
+    public Exam() {
     }
     
-     @Override
+    @Override
     public int hashCode() {
         int hash = 0;
         hash += (id != null ? id.hashCode() : 0);
@@ -118,10 +122,11 @@ public class Exercise implements Serializable{
     @Override
     public boolean equals(Object object) {
         // TODO: Warning - this method won't work in the case the id fields are not set
-        if (!(object instanceof Exercise)) {
+        if (!(object instanceof Exam)) {
             return false;
         }
-        Exercise other = (Exercise) object;
+        Exam other = (Exam) object;
         return !((this.id == null && other.id != null) || (this.id != null && !this.id.equals(other.id)));
     }
+
 }
