+++ src/java/entities/Mark.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package entitis;
+package entities;
 
 import java.io.File;
 import java.io.Serializable;
