+++ src/java/entities/Subject.java
@@ -1,4 +1,4 @@
-package entitis;
+package entities;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
